feat(urlTools) : expose all url tools and more in a dedicated module
